Update Section 3.4: particle counts to iu suffixes
The description that was there appeared to be for the old "particle counts", so I updated it for the new `iu` suffixes. You may want to further edit this section if you feel my adaptation does not make sense. E.g., you may feel that `iu` suffixes should be add _during_ the annotation process, which I could also see as reasonable.