compare_codings.py: new format (more compact, yet more informative)
If there are several messages of the same type for a given abstract and coder pair,
these will now be combined into a single block.
Furthermore, if sentence N had a problem,
sentence N+1 will be reported even if it does not.
This often avoids the need to open the abstract file for understanding the context.