Lutz adjusted Block 22 A and B
There was a curious message about non-aligned sentences near the end of CamRus22.txt.
It was due to a blank after the closing brace of a coding,
which SENTENCE_AND_ANNOTATION_PAIR_REGEXP cannot cope with.
Beware!
The fact that B was in CRLF encoding made this
much more difficult to find.